rename a header file.
